discord_bot/requirements.txt
discord.py==1.7.3
nltk==3.6.1
python-dotenv==0.19.2
gunicorn==20.1.0
tensorflow-cpu
